# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/scorecard/mock/preview": {
      "post": {
        "description": "Generates a mock scorecard for a given GitHub username and returns a frontend-ready UI model. Requires internal API key.",
        "operationId": "ScorecardController_preview",
        "parameters": [
          {
            "name": "X-Internal-Key",
            "in": "header",
            "description": "Internal API key (required)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewScorecardRequestDto"
              },
              "examples": {
                "example1": {
                  "summary": "Basic request",
                  "value": {
                    "githubUsername": "octocat"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully generated UI scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardUiDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid internal API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Preview scorecard (UI model)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/mock/preview/raw": {
      "post": {
        "description": "Returns full raw scoring data for debugging and internal analysis. Not intended for frontend usage.",
        "operationId": "ScorecardController_previewRaw",
        "parameters": [
          {
            "name": "X-Internal-Key",
            "in": "header",
            "description": "Internal API key (required)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewScorecardRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Raw scorecard data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardRawResponseDto"
                }
              }
            }
          }
        },
        "summary": "Preview scorecard (raw model)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/me": {
      "get": {
        "description": "Returns the authenticated user’s scorecard formatted for frontend display.",
        "operationId": "ScorecardController_getMyScorecard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardUiDto"
                }
              }
            }
          },
          "404": {
            "description": "No scorecard found. User must trigger GitHub sync first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my scorecard (UI)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/{username}": {
      "get": {
        "description": "Fetch a cached scorecard for a GitHub username. Returns frontend-ready UI model.",
        "operationId": "ScorecardController_getPublicScorecard",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "GitHub username",
            "schema": {
              "example": "octocat",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardUiDto"
                }
              }
            }
          },
          "404": {
            "description": "No cached scorecard found. Must trigger analysis first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public scorecard (UI)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/registered/{username}": {
      "get": {
        "description": "Fetch a cached scorecard for a registered username. Returns frontend-ready UI model.",
        "operationId": "ScorecardController_getPublicRegScorecard",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "registered username",
            "schema": {
              "example": "octocat",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardUiDto"
                }
              }
            }
          },
          "404": {
            "description": "No cached scorecard found. Must trigger analysis first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public scorecard (UI)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/me/raw": {
      "get": {
        "description": "Returns raw internal scorecard data for the authenticated user.",
        "operationId": "ScorecardController_getMyScorecardRaw",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Raw scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardRawResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my scorecard (raw)",
        "tags": [
          "Scorecard"
        ]
      }
    },
    "/api/scorecard/{username}/raw": {
      "get": {
        "description": "Returns raw cached scorecard for debugging or internal usage.",
        "operationId": "ScorecardController_getPublicScorecardRaw",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "example": "octocat",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw scorecard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardRawResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get public scorecard (raw)",
        "tags": [
          "Scorecard"
        ]
      }
    },
...
  },
  "info": {
    "title": "a16zero Employer API",
    "description": "Backend MVP for Employer platform features and Account Abstraction auth verification.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "internal-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-key",
        "description": "Internal API key"
      }
    },
    "schemas": {
...
      "PreviewScorecardRequestDto": {
        "type": "object",
        "properties": {}
      },
      "ScorecardUiDto": {
        "type": "object",
        "properties": {
          "profile": {
            "type": "object",
            "properties": {
              "username": {
                "type": "string",
                "description": "GitHub username of the candidate"
              },
              "avatarUrl": {
                "type": "string",
                "format": "uri"
              },
              "summary": {
                "type": "string"
              }
            },
            "required": [
              "username",
              "summary"
            ]
          },
          "score": {
            "type": "object",
            "properties": {
              "value": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "percentile": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "isWithheld": {
                "type": "object",
                "properties": {
                  "value": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "value"
                ]
              }
            },
            "required": [
              "value",
              "percentile",
              "isWithheld"
            ]
          },
          "trust": {
            "type": "object",
            "properties": {
              "level": {
                "type": "string"
              },
              "risk": {
                "type": "string",
                "enum": [
                  "LOW_RISK",
                  "MEDIUM_RISK",
                  "HIGH_RISK",
                  "INSUFFICIENT_DATA"
                ]
              },
              "label": {
                "type": "string"
              },
              "guidance": {
                "type": "string"
              }
            },
            "required": [
              "level",
              "risk",
              "label",
              "guidance"
            ]
          },
          "insights": {
            "type": "object",
            "properties": {
              "capabilities": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "key": {
                      "type": "string",
                      "enum": [
                        "backend",
                        "frontend",
                        "devops"
                      ]
                    },
                    "label": {
                      "type": "string"
                    },
                    "score": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 1
                    },
                    "displayScore": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 100
                    },
                    "confidence": {
                      "type": "string"
                    },
                    "strength": {
                      "type": "string",
                      "enum": [
                        "strong",
                        "moderate",
                        "weak"
                      ]
                    }
                  },
                  "required": [
                    "key",
                    "label",
                    "score",
                    "displayScore",
                    "confidence",
                    "strength"
                  ]
                }
              },
              "highlights": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "gaps": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "caveats": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "ownership": {
                "type": "object",
                "properties": {
                  "ownedProjects": {
                    "type": "number"
                  },
                  "activelyMaintained": {
                    "type": "number"
                  },
                  "confidence": {
                    "type": "string"
                  }
                },
                "required": [
                  "ownedProjects",
                  "activelyMaintained",
                  "confidence"
                ]
              },
              "impact": {
                "type": "object",
                "properties": {
                  "activityLevel": {
                    "type": "string"
                  },
                  "consistency": {
                    "type": "string"
                  },
                  "externalContributions": {
                    "type": "number"
                  },
                  "confidence": {
                    "type": "string"
                  }
                },
                "required": [
                  "activityLevel",
                  "consistency",
                  "externalContributions",
                  "confidence"
                ]
              },
              "reputation": {
                "nullable": true
              },
              "privateWorkNote": {
                "type": "string"
              },
              "organizations": {
                "type": "array",
                "items": {}
              },
              "interactionProfile": {
                "nullable": true
              },
              "stack": {
                "type": "object",
                "properties": {
                  "languages": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "tools": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "languages",
                  "tools"
                ]
              },
              "web3": {
                "nullable": true
              }
            },
            "required": [
              "capabilities",
              "highlights",
              "gaps",
              "caveats",
              "ownership",
              "impact"
            ]
          }
        },
        "required": [
          "profile",
          "score",
          "trust",
          "insights"
        ]
      },
      "ScorecardErrorResponseDto": {
        "type": "object",
        "properties": {}
      },
      "ScorecardRawResponseDto": {},
...
    }
  }
}